requirements.txt
arrow==0.13.1
asn1crypto==0.24.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bs4==0.0.1
bson==0.5.8
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
csvalidate==1.1.1
Django==2.2.4
djangorestframework==3.9.4
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
Flask==1.0.2
Flask-API==1.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.7
Flask-CSV==1.2.0
Flask-Login==0.4.1
Flask-SocketIO==3.3.2
Flask-Uploads==0.2.1
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gspread==3.1.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jinja2-time==0.2.0
jsonschema==2.6.0
MarkupSafe==1.1.1
marshmallow==2.19.2
nexmo==2.3.0
oauth2client==4.1.3
oauthlib==3.0.2
paramiko==2.4.2
pyasn1==0.4.6
pyasn1-modules==0.2.5
pycosat==0.6.3
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
python-engineio==3.9.0
python-socketio==3.1.2
pytz==2019.1
PyYAML==4.2b1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel-yaml==0.15.46
six==1.12.0
soupsieve==1.9.2
texttable==0.9.1
twilio==6.27.1
uritemplate==3.0.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.3
OdooRPC==0.7.0
odoo-rpc-client==1.0.0
odoo-autodiscover==2.0.0
odoo-import-export-client==2.2.0
constants==0.6.0
